backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
Flask-Cors
Flask-SocketIO
psycopg2-binary
python-dotenv
gunicorn
